Set file name of e2e timeout error screenshot
